requirements.txt
python-dotenv
fish-audio-sdk
flask
flask-cors
anthropic
instructor
